Fix CWE-190 / CWE-192 / CWE-197 / CWE-681

+++ NachoSpigot-API/src/main/java/co/aikar/timings/TimingData.java
@@ -59,7 +59,7 @@ class TimingData {
 
     void add(long diff) {
         ++curTickCount;
-        curTickTotal += diff;
+        curTickTotal += (int) diff;
     }
 
     void processTick(boolean violated) {
